Rename folder adapters
