+++ lib/features/fund_exchange/adapters/funding_gateway/models/get_funding_details_request_params_model.dart
