+++ lib/features/fund_exchange/adapters/funding_gateway/models/get_funding_details_response_model.dart
