+++ lib/features/fund_exchange/fund_exchange_locator.dart
@@ -5,9 +5,9 @@ import 'package:bb_mobile/features/fund_exchange/application/ports/funding_gatew
 import 'package:bb_mobile/features/fund_exchange/application/usecases/get_funding_details_usecase.dart';
 import 'package:bb_mobile/features/fund_exchange/application/usecases/list_funding_institutions_usecase.dart';
 import 'package:bb_mobile/features/fund_exchange/application/usecases/register_responsibility_consent_usecase.dart';
-import 'package:bb_mobile/features/fund_exchange/interface_adapters/settings_exchange_environment_adapter.dart';
-import 'package:bb_mobile/features/fund_exchange/interface_adapters/funding_gateway/bullbitcoin_api_funding_gateway.dart';
-import 'package:bb_mobile/features/fund_exchange/interface_adapters/funding_gateway/delegating_funding_gateway.dart';
+import 'package:bb_mobile/features/fund_exchange/adapters/settings_exchange_environment_adapter.dart';
+import 'package:bb_mobile/features/fund_exchange/adapters/funding_gateway/bullbitcoin_api_funding_gateway.dart';
+import 'package:bb_mobile/features/fund_exchange/adapters/funding_gateway/delegating_funding_gateway.dart';
 import 'package:bb_mobile/features/fund_exchange/presentation/bloc/fund_exchange_bloc.dart';
 import 'package:dio/dio.dart';
 import 'package:get_it/get_it.dart';
